+++ storagemarket/impl/clientstates/client_states_test.go
@@ -341,6 +341,7 @@ func TestCheckForDealAcceptance(t *testing.T) {
 				tut.AssertDealState(t, storagemarket.StorageDealCheckForAcceptance, deal.State)
 				assert.Equal(t, uint64(1), deal.PollRetryCount)
 				assert.Equal(t, uint64(1), deal.PollErrorCount)
+				assert.Equal(t, "Provider state: StorageDealUnknown", deal.Message)
 			},
 		})
 	})
@@ -352,6 +353,7 @@ func TestCheckForDealAcceptance(t *testing.T) {
 			},
 			inspector: func(deal storagemarket.ClientDeal, env *fakeEnvironment) {
 				tut.AssertDealState(t, storagemarket.StorageDealCheckForAcceptance, deal.State)
+				assert.Equal(t, "Provider state: StorageDealVerifyData", deal.Message)
 			},
 		})
 	})