+++ storagemarket/network/libp2p_impl_test.go
@@ -49,7 +49,7 @@ func TestOpenStreamWithRetries(t *testing.T) {
 	ctx := context.Background()
 	td := shared_testutil.NewLibp2pTestData(ctx, t)
 
-	fromNetwork := network.NewFromLibp2pHost(td.Host1, network.RetryParameters(1*time.Second, 10*time.Second, 5))
+	fromNetwork := network.NewFromLibp2pHost(td.Host1, network.RetryParameters(1*time.Second, 10*time.Second, 5, 5))
 	toNetwork := network.NewFromLibp2pHost(td.Host2)
 	toHost := td.Host2.ID()
 
@@ -531,7 +531,7 @@ func TestLibp2pStorageMarketNetwork_StopHandlingRequests(t *testing.T) {
 	bgCtx := context.Background()
 	td := shared_testutil.NewLibp2pTestData(bgCtx, t)
 
-	fromNetwork := network.NewFromLibp2pHost(td.Host1, network.RetryParameters(0, 0, 0))
+	fromNetwork := network.NewFromLibp2pHost(td.Host1, network.RetryParameters(0, 0, 0, 0))
 	toNetwork := network.NewFromLibp2pHost(td.Host2)
 	toHost := td.Host2.ID()