+++ storagemarket/testharness/testharness.go
@@ -43,12 +43,12 @@ type StorageHarness struct {
 func NewHarness(t *testing.T, ctx context.Context, useStore bool, cd testnodes.DelayFakeCommonNode, pd testnodes.DelayFakeCommonNode,
 	disableNewDeals bool) *StorageHarness {
 	smState := testnodes.NewStorageMarketState()
-	return NewHarnessWithTestData(t, ctx, shared_testutil.NewLibp2pTestData(ctx, t), smState, useStore, "", cd, pd, disableNewDeals)
+	td := shared_testutil.NewLibp2pTestData(ctx, t)
+	deps := dependencies.NewDependenciesWithTestData(t, ctx, td, smState, "", cd, pd)
+	return NewHarnessWithTestData(t, td, deps, useStore, disableNewDeals)
 }
 
-func NewHarnessWithTestData(t *testing.T, ctx context.Context, td *shared_testutil.Libp2pTestData, smState *testnodes.StorageMarketState, useStore bool, tempPath string,
-	cd testnodes.DelayFakeCommonNode, pd testnodes.DelayFakeCommonNode, disableNewDeals bool) *StorageHarness {
-	deps := dependencies.NewDependenciesWithTestData(t, ctx, td, smState, tempPath, cd, pd)
+func NewHarnessWithTestData(t *testing.T, td *shared_testutil.Libp2pTestData, deps *dependencies.StorageDependencies, useStore bool, disableNewDeals bool) *StorageHarness {
 	fpath := filepath.Join("storagemarket", "fixtures", "payload.txt")
 	var rootLink ipld.Link
 	var storeID *multistore.StoreID
@@ -65,7 +65,7 @@ func NewHarnessWithTestData(t *testing.T, ctx context.Context, td *shared_testut
 
 	clientDs := namespace.Wrap(td.Ds1, datastore.NewKey("/deals/client"))
 	client, err := storageimpl.NewClient(
-		network.NewFromLibp2pHost(td.Host1, network.RetryParameters(0, 0, 0)),
+		network.NewFromLibp2pHost(td.Host1, network.RetryParameters(0, 0, 0, 0)),
 		td.Bs1,
 		td.MultiStore1,
 		deps.DTClient,
@@ -77,7 +77,7 @@ func NewHarnessWithTestData(t *testing.T, ctx context.Context, td *shared_testut
 	require.NoError(t, err)
 
 	providerDs := namespace.Wrap(td.Ds1, datastore.NewKey("/deals/provider"))
-	networkOptions := []network.Option{network.RetryParameters(0, 0, 0)}
+	networkOptions := []network.Option{network.RetryParameters(0, 0, 0, 0)}
 	/* if disableNewDeals {
 		networkOptions = append(networkOptions,
 			network.SupportedAskProtocols([]protocol.ID{storagemarket.OldAskProtocolID}),
@@ -111,16 +111,16 @@ func NewHarnessWithTestData(t *testing.T, ctx context.Context, td *shared_testut
 	}
 }
 
-func (h *StorageHarness) CreateNewProvider(t *testing.T, ctx context.Context, td *shared_testutil.Libp2pTestData, tempPath string, disableNewDeals bool) {
+func (h *StorageHarness) CreateNewProvider(t *testing.T, ctx context.Context, td *shared_testutil.Libp2pTestData) storagemarket.StorageProvider {
 	gs2 := graphsyncimpl.New(ctx, gsnetwork.NewFromLibp2pHost(td.Host2), td.Loader2, td.Storer2)
 	dtTransport2 := dtgstransport.NewTransport(td.Host2.ID(), gs2)
-	dt2, err := dtimpl.NewDataTransfer(td.DTStore2, td.DTNet2, dtTransport2, td.DTStoredCounter2)
+	dt2, err := dtimpl.NewDataTransfer(td.DTStore2, td.DTTmpDir2, td.DTNet2, dtTransport2, td.DTStoredCounter2)
 	require.NoError(t, err)
 	testutil.StartAndWaitForReady(ctx, t, dt2)
 
 	providerDs := namespace.Wrap(td.Ds1, datastore.NewKey("/deals/provider"))
 	provider, err := storageimpl.NewProvider(
-		network.NewFromLibp2pHost(td.Host2, network.RetryParameters(0, 0, 0)),
+		network.NewFromLibp2pHost(td.Host2, network.RetryParameters(0, 0, 0, 0)),
 		providerDs,
 		h.Fs,
 		td.MultiStore2,
@@ -131,7 +131,7 @@ func (h *StorageHarness) CreateNewProvider(t *testing.T, ctx context.Context, td
 		h.StoredAsk,
 	)
 	require.NoError(t, err)
-	h.Provider = provider
+	return provider
 }
 
 func (h *StorageHarness) ProposeStorageDeal(t *testing.T, dataRef *storagemarket.DataRef, fastRetrieval /* , verifiedDeal */ bool) *storagemarket.ProposeStorageDealResult {